Multitudes of issue fixes and back-end features have been added.
-Audio classes now has a linearly scaling volume operator.
-"Smart" audio system has been added and is enforced. Music tracks must
now be preloaded before playing.
-Main menu scaling bug has been fixed.
-Rendering cycles have been expanded on and improved. There now exists
EarlyUpdates, normal Updates, and GraphicUpdates (called in that order).
LateUpdates will added in the future.
-Implemented a Time class which will deal with all functions of time
instead of being calculated in the main render method.
-Main render method has been reworked and is much more lightweight.
-A lot of functionality in the gameplay is currently being rewritten to
use the new Update system.
-A removal list has been added for entities. Whenever an entity needs to
be removed, it will get added to the removal list and removed at the end
of the frame. This fixes the next issue:
-Aliens in pack no longer blink out and in to existence when an adjacent
alien has been shot.
-Reintroduced and re-removed the issue of being able to kill multiple
aliens with one shot.

-The game is currently inoperable due to the WIP layered rendering
system. All things, including entities, which need to be drawn, will all
be drawn using the standard GraphicsUpdate method. This requires
rewrites in the base Entity class, along with any previously working
GUI.