Logger overhaul and major file fix. (0.pre_3.0 | 9-6.27.18)
-The Logger class now has a method to send exceptions to so they can be
properly logged.
-The Logger class now properly catches uncaught exceptions.
-The Logger class has a new boolean to toggle whether or not only errors
should be printed to the output file.
-The Logger class will now only create a file if the printToFile is true
and if something will be written to it.
-In the Logger class, printToFile and onlyPrintErrsToFile are both
defaulted to true. This means that output files will be created in the
development workspace if there is an error during runtime.
-The space background image file has been renamed to "spaceBG1.png" (was
previously "SpaceBG1.PNG") to match the references to it in the code.
-Moved where the version numbers are stored to the ConstantValues class.

After fixing the image issue, exporting runnable jars now works
properly.
After overhauling the Logger, the logging system now works as intended.